requirements.txt
llama-index
llama-index-readers-file
llama-index-vector-stores-faiss
llama-index-retrievers-bm25
openai
anthropic
faiss-cpu
llama-index-embeddings-openai
llama-index-llms-openai
pillow>=10.4.0
beautifulsoup4
nest-asyncio
boto3
python-dotenv
google-genai
pyyaml
pydantic>=2.0.0
pydantic-settings
huggingface-hub>=0.23.0
